Implement Client Agent Auto Config
There are a couple of things in here.

First, just like auto encrypt, any Cluster.AutoConfig RPC will implicitly use the less secure RPC mechanism.

Secondly there are some oddities with how an agent starts up. Half being in the command/agent package and half being in the agent package. Previously all configuration loading was done in the command/agent package and a config was presented to the Agent. Things are more complicated now with the agent needing to inject other configuration that the command package doesn’t know about. However the command package still needs to parse config prior to creating the agent and then use some configurations independent of the agent. So to accomplish this  the command first parses the config and passes it along with the cli flags to the agent. If auto config is enabled the agent will get the auto-config and then rebuild the configuration taking it into account. Later on when responding to reload requests (which happen in the cli file) it will use a method on the agent to load the config so that we don’t inadvertently forget about the auto-config.

Honestly this is pretty messy and we should refactor things at some point to take more of the logic out of the CLI command and put it in the agent. That is a task for the future though.

fixup rpc